// frontend/src/sections/patients/list/insertPatient.generated.ts
/* eslint-disable */
// THIS FILE IS GENERATED by codegen/plugin.js — do not edit by hand.
import type { TypedDocument } from "../../../api/graphql";

export type InsertPatientVariables = {
  storeId: string;
  input: {
    id: string;
    code: string;
    code2?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    gender?: "FEMALE" | "MALE" | "TRANSGENDER" | "TRANSGENDER_MALE" | "TRANSGENDER_MALE_HORMONE" | "TRANSGENDER_MALE_SURGICAL" | "TRANSGENDER_FEMALE" | "TRANSGENDER_FEMALE_HORMONE" | "TRANSGENDER_FEMALE_SURGICAL" | "UNKNOWN" | "NON_BINARY" | null;
    dateOfBirth?: string | null;
    address1?: string | null;
    phone?: string | null;
    isDeceased?: boolean | null;
    dateOfDeath?: string | null;
    nextOfKinId?: string | null;
    nextOfKinName?: string | null;
  };
};

export type InsertPatientResult = {
  insertPatient: ({
  __typename: "PatientNode";
} & {
  id: string;
});
};

export const InsertPatient = {
  query: "mutation insertPatient($storeId: String!, $input: InsertPatientInput!) {\n  insertPatient(storeId: $storeId, input: $input) {\n    ... on PatientNode {\n      __typename\n      id\n    }\n  }\n}",
} as TypedDocument<InsertPatientResult, InsertPatientVariables>;